Introducing runtime version-specific checks
